fix: parse markdown 로직 분리
